@@ -1,5 +1,5 @@
-"""Tests for main driver for price fetching.
-"""
+"""Tests for main driver for price fetching."""
+
 __copyright__ = "Copyright (C) 2015-2020  Martin Blais"
 __license__ = "GNU GPLv2"
 
@@ -56,36 +56,36 @@ def run_with_args(function, args, runner_file=None):
 
 
 class TestCache(unittest.TestCase):
-
     def test_fetch_cached_price__disabled(self):
         # Latest.
-        with mock.patch('beanprice.price._CACHE', None):
+        with mock.patch("beanprice.price._CACHE", None):
             self.assertIsNone(price._CACHE)
             source = mock.MagicMock()
-            price.fetch_cached_price(source, 'HOOL', None)
+            price.fetch_cached_price(source, "HOOL", None)
             self.assertTrue(source.get_latest_price.called)
 
         # Historical.
-        with mock.patch('beanprice.price._CACHE', None):
+        with mock.patch("beanprice.price._CACHE", None):
             self.assertIsNone(price._CACHE)
             source = mock.MagicMock()
-            price.fetch_cached_price(source, 'HOOL', datetime.date.today())
+            price.fetch_cached_price(source, "HOOL", datetime.date.today())
             self.assertTrue(source.get_historical_price.called)
 
     def test_fetch_cached_price__latest(self):
         tmpdir = tempfile.mkdtemp()
-        tmpfile = path.join(tmpdir, 'prices.cache')
+        tmpfile = path.join(tmpdir, "prices.cache")
         try:
             price.setup_cache(tmpfile, False)
 
-            srcprice = SourcePrice(Decimal('1.723'), datetime.datetime.now(tz.tzutc()),
-                                   'USD')
+            srcprice = SourcePrice(
+                Decimal("1.723"), datetime.datetime.now(tz.tzutc()), "USD"
+            )
             source = mock.MagicMock()
             source.get_latest_price.return_value = srcprice
-            source.__file__ = '<module>'
+            source.__file__ = "<module>"
 
             # Cache miss.
-            result = price.fetch_cached_price(source, 'HOOL', None)
+            result = price.fetch_cached_price(source, "HOOL", None)
             self.assertTrue(source.get_latest_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice, result)
@@ -93,20 +93,21 @@ class TestCache(unittest.TestCase):
             source.get_latest_price.reset_mock()
 
             # Cache hit.
-            result = price.fetch_cached_price(source, 'HOOL', None)
+            result = price.fetch_cached_price(source, "HOOL", None)
             self.assertFalse(source.get_latest_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice, result)
 
             srcprice2 = SourcePrice(
-                Decimal('1.894'), datetime.datetime.now(tz.tzutc()), 'USD')
+                Decimal("1.894"), datetime.datetime.now(tz.tzutc()), "USD"
+            )
             source.get_latest_price.reset_mock()
             source.get_latest_price.return_value = srcprice2
 
             # Cache expired.
             time_beyond = datetime.datetime.now() + price._CACHE.expiration * 2
-            with mock.patch('beanprice.price.now', return_value=time_beyond):
-                result = price.fetch_cached_price(source, 'HOOL', None)
+            with mock.patch("beanprice.price.now", return_value=time_beyond):
+                result = price.fetch_cached_price(source, "HOOL", None)
                 self.assertTrue(source.get_latest_price.called)
                 self.assertEqual(1, len(price._CACHE))
                 self.assertEqual(srcprice2, result)
@@ -117,18 +118,19 @@ class TestCache(unittest.TestCase):
 
     def test_fetch_cached_price__clear_cache(self):
         tmpdir = tempfile.mkdtemp()
-        tmpfile = path.join(tmpdir, 'prices.cache')
+        tmpfile = path.join(tmpdir, "prices.cache")
         try:
             price.setup_cache(tmpfile, False)
 
-            srcprice = SourcePrice(Decimal('1.723'), datetime.datetime.now(tz.tzutc()),
-                                   'USD')
+            srcprice = SourcePrice(
+                Decimal("1.723"), datetime.datetime.now(tz.tzutc()), "USD"
+            )
             source = mock.MagicMock()
             source.get_latest_price.return_value = srcprice
-            source.__file__ = '<module>'
+            source.__file__ = "<module>"
 
             # Cache miss.
-            result = price.fetch_cached_price(source, 'HOOL', None)
+            result = price.fetch_cached_price(source, "HOOL", None)
             self.assertTrue(source.get_latest_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice, result)
@@ -136,13 +138,14 @@ class TestCache(unittest.TestCase):
             source.get_latest_price.reset_mock()
 
             # Cache hit.
-            result = price.fetch_cached_price(source, 'HOOL', None)
+            result = price.fetch_cached_price(source, "HOOL", None)
             self.assertFalse(source.get_latest_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice, result)
 
             srcprice2 = SourcePrice(
-                Decimal('1.894'), datetime.datetime.now(tz.tzutc()), 'USD')
+                Decimal("1.894"), datetime.datetime.now(tz.tzutc()), "USD"
+            )
             source.get_latest_price.reset_mock()
             source.get_latest_price.return_value = srcprice2
 
@@ -151,7 +154,7 @@ class TestCache(unittest.TestCase):
             price.setup_cache(tmpfile, True)
 
             # Cache cleared.
-            result = price.fetch_cached_price(source, 'HOOL', None)
+            result = price.fetch_cached_price(source, "HOOL", None)
             self.assertTrue(source.get_latest_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice2, result)
@@ -162,19 +165,20 @@ class TestCache(unittest.TestCase):
 
     def test_fetch_cached_price__historical(self):
         tmpdir = tempfile.mkdtemp()
-        tmpfile = path.join(tmpdir, 'prices.cache')
+        tmpfile = path.join(tmpdir, "prices.cache")
         try:
             price.setup_cache(tmpfile, False)
 
             srcprice = SourcePrice(
-                Decimal('1.723'), datetime.datetime.now(tz.tzutc()), 'USD')
+                Decimal("1.723"), datetime.datetime.now(tz.tzutc()), "USD"
+            )
             source = mock.MagicMock()
             source.get_historical_price.return_value = srcprice
-            source.__file__ = '<module>'
+            source.__file__ = "<module>"
 
             # Cache miss.
             day = datetime.date(2006, 1, 2)
-            result = price.fetch_cached_price(source, 'HOOL', day)
+            result = price.fetch_cached_price(source, "HOOL", day)
             self.assertTrue(source.get_historical_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice, result)
@@ -182,7 +186,7 @@ class TestCache(unittest.TestCase):
             source.get_historical_price.reset_mock()
 
             # Cache hit.
-            result = price.fetch_cached_price(source, 'HOOL', day)
+            result = price.fetch_cached_price(source, "HOOL", day)
             self.assertFalse(source.get_historical_price.called)
             self.assertEqual(1, len(price._CACHE))
             self.assertEqual(srcprice, result)
@@ -193,12 +197,10 @@ class TestCache(unittest.TestCase):
 
 
 class TestProcessArguments(unittest.TestCase):
-
     def test_filename_not_exists(self):
-        with test_utils.capture('stderr'):
+        with test_utils.capture("stderr"):
             with self.assertRaises(SystemExit):
-                run_with_args(
-                    price.process_args, ['--no-cache', '/some/file.beancount'])
+                run_with_args(price.process_args, ["--no-cache", "/some/file.beancount"])
 
     @test_utils.docfile
     def test_explicit_file__badcontents(self, filename):
@@ -206,49 +208,54 @@ class TestProcessArguments(unittest.TestCase):
         2015-01-01 open Assets:Invest
         2015-01-01 open USD ;; Error
         """
-        with test_utils.capture('stderr'):
-            args, jobs, _, __ = run_with_args(
-                price.process_args, ['--no-cache', filename])
+        with test_utils.capture("stderr"):
+            args, jobs, _, __ = run_with_args(price.process_args, ["--no-cache", filename])
             self.assertEqual([], jobs)
 
     def test_filename_exists(self):
-        with tempfile.NamedTemporaryFile('w') as tmpfile:
-            with test_utils.capture('stderr'):
+        with tempfile.NamedTemporaryFile("w") as tmpfile:
+            with test_utils.capture("stderr"):
                 args, jobs, _, __ = run_with_args(
-                    price.process_args, ['--no-cache', tmpfile.name])
+                    price.process_args, ["--no-cache", tmpfile.name]
+                )
                 self.assertEqual([], jobs)  # Empty file.
 
     def test_expressions(self):
-        with test_utils.capture('stderr'):
+        with test_utils.capture("stderr"):
             args, jobs, _, __ = run_with_args(
-                price.process_args, ['--no-cache', '-e', 'USD:yahoo/AAPL'])
+                price.process_args, ["--no-cache", "-e", "USD:yahoo/AAPL"]
+            )
             self.assertEqual(
-                [price.DatedPrice(
-                    'AAPL', 'USD', None,
-                    [price.PriceSource(yahoo, 'AAPL', False)])], jobs)
+                [
+                    price.DatedPrice(
+                        "AAPL", "USD", None, [price.PriceSource(yahoo, "AAPL", False)]
+                    )
+                ],
+                jobs,
+            )
 
 
 class TestClobber(cmptest.TestCase):
-
     @loader.load_doc()
     def setUp(self, entries, _, __):
         """
-          ;; Existing file.
-          2015-01-05 price HDV                                 75.56 USD
-          2015-01-23 price HDV                                 77.34 USD
-          2015-02-06 price HDV                                 77.16 USD
-          2015-02-12 price HDV                                 78.17 USD
-          2015-05-01 price HDV                                 77.48 USD
-          2015-06-02 price HDV                                 76.33 USD
-          2015-06-29 price HDV                                 73.74 USD
-          2015-07-06 price HDV                                 73.79 USD
-          2015-08-11 price HDV                                 74.19 USD
-          2015-09-04 price HDV                                 68.98 USD
+        ;; Existing file.
+        2015-01-05 price HDV                                 75.56 USD
+        2015-01-23 price HDV                                 77.34 USD
+        2015-02-06 price HDV                                 77.16 USD
+        2015-02-12 price HDV                                 78.17 USD
+        2015-05-01 price HDV                                 77.48 USD
+        2015-06-02 price HDV                                 76.33 USD
+        2015-06-29 price HDV                                 73.74 USD
+        2015-07-06 price HDV                                 73.79 USD
+        2015-08-11 price HDV                                 74.19 USD
+        2015-09-04 price HDV                                 68.98 USD
         """
         self.entries = entries
 
         # New entries.
-        self.price_entries, _, __ = loader.load_string("""
+        self.price_entries, _, __ = loader.load_string(
+            """
           2015-01-27 price HDV                                 76.83 USD
           2015-02-06 price HDV                                 77.16 USD
           2015-02-19 price HDV                                  77.5 USD
@@ -257,163 +264,185 @@ class TestClobber(cmptest.TestCase):
           2015-07-06 price HDV                                 73.79 USD
           2015-07-31 price HDV                                 74.64 USD
           2015-08-11 price HDV                                 74.20 USD ;; Different
-        """, dedent=True)
+        """,
+            dedent=True,
+        )
 
     def test_clobber_nodiffs(self):
-        new_price_entries, _ = price.filter_redundant_prices(self.price_entries,
-                                                             self.entries,
-                                                             diffs=False)
-        self.assertEqualEntries("""
+        new_price_entries, _ = price.filter_redundant_prices(
+            self.price_entries, self.entries, diffs=False
+        )
+        self.assertEqualEntries(
+            """
           2015-01-27 price HDV                                 76.83 USD
           2015-02-19 price HDV                                  77.5 USD
           2015-06-19 price HDV                                    76 USD
           2015-07-31 price HDV                                 74.64 USD
-        """, new_price_entries)
+        """,
+            new_price_entries,
+        )
 
     def test_clobber_diffs(self):
-        new_price_entries, _ = price.filter_redundant_prices(self.price_entries,
-                                                             self.entries,
-                                                             diffs=True)
-        self.assertEqualEntries("""
+        new_price_entries, _ = price.filter_redundant_prices(
+            self.price_entries, self.entries, diffs=True
+        )
+        self.assertEqualEntries(
+            """
           2015-01-27 price HDV                                 76.83 USD
           2015-02-19 price HDV                                  77.5 USD
           2015-06-19 price HDV                                    76 USD
           2015-07-31 price HDV                                 74.64 USD
           2015-08-11 price HDV                                 74.20 USD ;; Different
-        """, new_price_entries)
+        """,
+            new_price_entries,
+        )
 
 
 class TestTimezone(unittest.TestCase):
-
-    @mock.patch.object(price, 'fetch_cached_price')
+    @mock.patch.object(price, "fetch_cached_price")
     def test_fetch_price__naive_time_no_timeozne(self, fetch_cached):
         fetch_cached.return_value = SourcePrice(
-            Decimal('125.00'), datetime.datetime(2015, 11, 22, 16, 0, 0), 'JPY')
-        dprice = price.DatedPrice('JPY', 'USD', datetime.date(2015, 11, 22), None)
+            Decimal("125.00"), datetime.datetime(2015, 11, 22, 16, 0, 0), "JPY"
+        )
+        dprice = price.DatedPrice("JPY", "USD", datetime.date(2015, 11, 22), None)
         with self.assertRaises(ValueError):
-            price.fetch_price(dprice._replace(sources=[
-                price.PriceSource(yahoo, 'USDJPY', False)]), False)
+            price.fetch_price(
+                dprice._replace(sources=[price.PriceSource(yahoo, "USDJPY", False)]), False
+            )
 
 
 class TestInverted(unittest.TestCase):
-
     def setUp(self):
-        fetch_cached = mock.patch('beanprice.price.fetch_cached_price').start()
+        fetch_cached = mock.patch("beanprice.price.fetch_cached_price").start()
         fetch_cached.return_value = SourcePrice(
-            Decimal('125.00'), datetime.datetime(2015, 11, 22, 16, 0, 0,
-                                                 tzinfo=tz.tzlocal()),
-            'JPY')
-        self.dprice = price.DatedPrice('JPY', 'USD', datetime.date(2015, 11, 22),
-                                       None)
+            Decimal("125.00"),
+            datetime.datetime(2015, 11, 22, 16, 0, 0, tzinfo=tz.tzlocal()),
+            "JPY",
+        )
+        self.dprice = price.DatedPrice("JPY", "USD", datetime.date(2015, 11, 22), None)
         self.addCleanup(mock.patch.stopall)
 
     def test_fetch_price__normal(self):
-        entry = price.fetch_price(self.dprice._replace(sources=[
-            price.PriceSource(yahoo, 'USDJPY', False)]), False)
-        self.assertEqual(('JPY', 'USD'), (entry.currency, entry.amount.currency))
-        self.assertEqual(Decimal('125.00'), entry.amount.number)
+        entry = price.fetch_price(
+            self.dprice._replace(sources=[price.PriceSource(yahoo, "USDJPY", False)]), False
+        )
+        self.assertEqual(("JPY", "USD"), (entry.currency, entry.amount.currency))
+        self.assertEqual(Decimal("125.00"), entry.amount.number)
 
     def test_fetch_price__inverted(self):
-        entry = price.fetch_price(self.dprice._replace(sources=[
-            price.PriceSource(yahoo, 'USDJPY', True)]), False)
-        self.assertEqual(('JPY', 'USD'), (entry.currency, entry.amount.currency))
-        self.assertEqual(Decimal('0.008'), entry.amount.number)
+        entry = price.fetch_price(
+            self.dprice._replace(sources=[price.PriceSource(yahoo, "USDJPY", True)]), False
+        )
+        self.assertEqual(("JPY", "USD"), (entry.currency, entry.amount.currency))
+        self.assertEqual(Decimal("0.008"), entry.amount.number)
 
     def test_fetch_price__swapped(self):
-        entry = price.fetch_price(self.dprice._replace(sources=[
-            price.PriceSource(yahoo, 'USDJPY', True)]), True)
-        self.assertEqual(('USD', 'JPY'), (entry.currency, entry.amount.currency))
-        self.assertEqual(Decimal('125.00'), entry.amount.number)
+        entry = price.fetch_price(
+            self.dprice._replace(sources=[price.PriceSource(yahoo, "USDJPY", True)]), True
+        )
+        self.assertEqual(("USD", "JPY"), (entry.currency, entry.amount.currency))
+        self.assertEqual(Decimal("125.00"), entry.amount.number)
 
 
 class TestImportSource(unittest.TestCase):
-
     def test_import_source_valid(self):
-        for name in 'oanda', 'yahoo':
+        for name in "oanda", "yahoo":
             module = price.import_source(name)
             self.assertIsInstance(module, types.ModuleType)
-        module = price.import_source('beanprice.sources.yahoo')
+        module = price.import_source("beanprice.sources.yahoo")
         self.assertIsInstance(module, types.ModuleType)
 
     def test_import_source_invalid(self):
         with self.assertRaises(ImportError):
-            price.import_source('non.existing.module')
+            price.import_source("non.existing.module")
 
 
 class TestParseSource(unittest.TestCase):
-
     def test_source_invalid(self):
         with self.assertRaises(ValueError):
-            price.parse_single_source('AAPL')
+            price.parse_single_source("AAPL")
         with self.assertRaises(ValueError):
-            price.parse_single_source('***//--')
+            price.parse_single_source("***//--")
 
         # The module gets imported at this stage.
         with self.assertRaises(ImportError):
-            price.parse_single_source('invalid.module.name/NASDAQ:AAPL')
+            price.parse_single_source("invalid.module.name/NASDAQ:AAPL")
 
     def test_source_valid(self):
-        psource = price.parse_single_source('yahoo/CNYUSD=X')
-        self.assertEqual(PS(yahoo, 'CNYUSD=X', False), psource)
+        psource = price.parse_single_source("yahoo/CNYUSD=X")
+        self.assertEqual(PS(yahoo, "CNYUSD=X", False), psource)
 
         # Make sure that an invalid name at the tail doesn't succeed.
         with self.assertRaises(ValueError):
-            psource = price.parse_single_source('yahoo/CNYUSD&X')
+            psource = price.parse_single_source("yahoo/CNYUSD&X")
 
-        psource = price.parse_single_source('beanprice.sources.yahoo/AAPL')
-        self.assertEqual(PS(yahoo, 'AAPL', False), psource)
+        psource = price.parse_single_source("beanprice.sources.yahoo/AAPL")
+        self.assertEqual(PS(yahoo, "AAPL", False), psource)
 
 
 class TestParseSourceMap(unittest.TestCase):
-
     def _clean_source_map(self, smap):
-        return {currency: [PS(s[0].__name__, s[1], s[2]) for s in sources]
-                for currency, sources in smap.items()}
+        return {
+            currency: [PS(s[0].__name__, s[1], s[2]) for s in sources]
+            for currency, sources in smap.items()
+        }
 
     def test_source_map_invalid(self):
-        for expr in 'USD', 'something else', 'USD:NASDAQ:AAPL':
+        for expr in "USD", "something else", "USD:NASDAQ:AAPL":
             with self.assertRaises(ValueError):
                 price.parse_source_map(expr)
 
     def test_source_map_onecur_single(self):
-        smap = price.parse_source_map('USD:yahoo/AAPL')
+        smap = price.parse_source_map("USD:yahoo/AAPL")
         self.assertEqual(
-            {'USD': [PS('beanprice.sources.yahoo', 'AAPL', False)]},
-            self._clean_source_map(smap))
+            {"USD": [PS("beanprice.sources.yahoo", "AAPL", False)]},
+            self._clean_source_map(smap),
+        )
 
     def test_source_map_onecur_multiple(self):
-        smap = price.parse_source_map('USD:oanda/USDCAD,yahoo/CAD=X')
+        smap = price.parse_source_map("USD:oanda/USDCAD,yahoo/CAD=X")
         self.assertEqual(
-            {'USD': [PS('beanprice.sources.oanda', 'USDCAD', False),
-                     PS('beanprice.sources.yahoo', 'CAD=X', False)]},
-            self._clean_source_map(smap))
+            {
+                "USD": [
+                    PS("beanprice.sources.oanda", "USDCAD", False),
+                    PS("beanprice.sources.yahoo", "CAD=X", False),
+                ]
+            },
+            self._clean_source_map(smap),
+        )
 
     def test_source_map_manycur_single(self):
-        smap = price.parse_source_map('USD:yahoo/USDCAD '
-                                      'CAD:yahoo/CAD=X')
+        smap = price.parse_source_map("USD:yahoo/USDCAD CAD:yahoo/CAD=X")
         self.assertEqual(
-            {'USD': [PS('beanprice.sources.yahoo', 'USDCAD', False)],
-             'CAD': [PS('beanprice.sources.yahoo', 'CAD=X', False)]},
-            self._clean_source_map(smap))
+            {
+                "USD": [PS("beanprice.sources.yahoo", "USDCAD", False)],
+                "CAD": [PS("beanprice.sources.yahoo", "CAD=X", False)],
+            },
+            self._clean_source_map(smap),
+        )
 
     def test_source_map_manycur_multiple(self):
-        smap = price.parse_source_map('USD:yahoo/GBPUSD,oanda/GBPUSD '
-                                      'CAD:yahoo/GBPCAD')
+        smap = price.parse_source_map("USD:yahoo/GBPUSD,oanda/GBPUSD CAD:yahoo/GBPCAD")
         self.assertEqual(
-            {'USD': [PS('beanprice.sources.yahoo', 'GBPUSD', False),
-                     PS('beanprice.sources.oanda', 'GBPUSD', False)],
-             'CAD': [PS('beanprice.sources.yahoo', 'GBPCAD', False)]},
-            self._clean_source_map(smap))
+            {
+                "USD": [
+                    PS("beanprice.sources.yahoo", "GBPUSD", False),
+                    PS("beanprice.sources.oanda", "GBPUSD", False),
+                ],
+                "CAD": [PS("beanprice.sources.yahoo", "GBPCAD", False)],
+            },
+            self._clean_source_map(smap),
+        )
 
     def test_source_map_inverse(self):
-        smap = price.parse_source_map('USD:yahoo/^GBPUSD')
+        smap = price.parse_source_map("USD:yahoo/^GBPUSD")
         self.assertEqual(
-            {'USD': [PS('beanprice.sources.yahoo', 'GBPUSD', True)]},
-            self._clean_source_map(smap))
+            {"USD": [PS("beanprice.sources.yahoo", "GBPUSD", True)]},
+            self._clean_source_map(smap),
+        )
 
 
 class TestFilters(unittest.TestCase):
-
     @loader.load_doc()
     def test_get_price_jobs__date(self, entries, _, __):
         """
@@ -441,22 +470,21 @@ class TestFilters(unittest.TestCase):
           Assets:US:Invest:VEA            -200 VEA {43.22 USD} @ 41.01 USD
           Assets:US:Invest:Margin
         """
-        jobs = price.get_price_jobs_at_date(entries, datetime.date(2014, 1, 1),
-                                            False, None)
+        jobs = price.get_price_jobs_at_date(entries, datetime.date(2014, 1, 1), False, None)
         self.assertEqual(set(), {(job.base, job.quote) for job in jobs})
 
-        jobs = price.get_price_jobs_at_date(entries, datetime.date(2014, 6, 1),
-                                            False, None)
-        self.assertEqual({('QQQ', 'USD'), ('VEA', 'USD')},
-                         {(job.base, job.quote) for job in jobs})
+        jobs = price.get_price_jobs_at_date(entries, datetime.date(2014, 6, 1), False, None)
+        self.assertEqual(
+            {("QQQ", "USD"), ("VEA", "USD")}, {(job.base, job.quote) for job in jobs}
+        )
 
-        jobs = price.get_price_jobs_at_date(entries, datetime.date(2014, 10, 1),
-                                            False, None)
-        self.assertEqual({('VEA', 'USD')},
-                         {(job.base, job.quote) for job in jobs})
+        jobs = price.get_price_jobs_at_date(
+            entries, datetime.date(2014, 10, 1), False, None
+        )
+        self.assertEqual({("VEA", "USD")}, {(job.base, job.quote) for job in jobs})
 
         jobs = price.get_price_jobs_at_date(entries, None, False, None)
-        self.assertEqual({('QQQ', 'USD')}, {(job.base, job.quote) for job in jobs})
+        self.assertEqual({("QQQ", "USD")}, {(job.base, job.quote) for job in jobs})
 
     @loader.load_doc()
     def test_get_price_jobs__inactive(self, entries, _, __):
@@ -481,11 +509,12 @@ class TestFilters(unittest.TestCase):
           Assets:US:Invest:Margin
         """
         jobs = price.get_price_jobs_at_date(entries, None, False, None)
-        self.assertEqual({('VEA', 'USD')}, {(job.base, job.quote) for job in jobs})
+        self.assertEqual({("VEA", "USD")}, {(job.base, job.quote) for job in jobs})
 
         jobs = price.get_price_jobs_at_date(entries, None, True, None)
-        self.assertEqual({('VEA', 'USD'), ('QQQ', 'USD')},
-                         {(job.base, job.quote) for job in jobs})
+        self.assertEqual(
+            {("VEA", "USD"), ("QQQ", "USD")}, {(job.base, job.quote) for job in jobs}
+        )
 
     @loader.load_doc()
     def test_get_price_jobs__undeclared(self, entries, _, __):
@@ -503,11 +532,12 @@ class TestFilters(unittest.TestCase):
           Assets:US:Invest:Margin
         """
         jobs = price.get_price_jobs_at_date(entries, None, False, None)
-        self.assertEqual({('QQQ', 'USD')}, {(job.base, job.quote) for job in jobs})
+        self.assertEqual({("QQQ", "USD")}, {(job.base, job.quote) for job in jobs})
 
-        jobs = price.get_price_jobs_at_date(entries, None, False, 'yahoo')
-        self.assertEqual({('QQQ', 'USD'), ('VEA', 'USD')},
-                         {(job.base, job.quote) for job in jobs})
+        jobs = price.get_price_jobs_at_date(entries, None, False, "yahoo")
+        self.assertEqual(
+            {("QQQ", "USD"), ("VEA", "USD")}, {(job.base, job.quote) for job in jobs}
+        )
 
     @loader.load_doc()
     def test_get_price_jobs__default_source(self, entries, _, __):
@@ -522,7 +552,7 @@ class TestFilters(unittest.TestCase):
           Assets:US:Invest:QQQ             100 QQQ {86.23 USD}
           Assets:US:Invest:Margin
         """
-        jobs = price.get_price_jobs_at_date(entries, None, False, 'yahoo')
+        jobs = price.get_price_jobs_at_date(entries, None, False, "yahoo")
         self.assertEqual(1, len(jobs[0].sources))
         self.assertIsInstance(jobs[0].sources[0], price.PriceSource)
 
@@ -541,17 +571,16 @@ class TestFilters(unittest.TestCase):
           Assets:US:BofA:CHF             -110 CHF @@ 100 USD
         """
         # TODO: Shouldn't we actually return (CHF, USD) here?
-        jobs = price.get_price_jobs_at_date(entries, datetime.date(2021, 1, 4),
-                                            False, None)
+        jobs = price.get_price_jobs_at_date(entries, datetime.date(2021, 1, 4), False, None)
         self.assertEqual(set(), {(job.base, job.quote) for job in jobs})
 
-        jobs = price.get_price_jobs_at_date(entries, datetime.date(2021, 1, 6),
-                                            False, None)
-        self.assertEqual({('CHF', 'USD')}, {(job.base, job.quote) for job in jobs})
+        jobs = price.get_price_jobs_at_date(entries, datetime.date(2021, 1, 6), False, None)
+        self.assertEqual({("CHF", "USD")}, {(job.base, job.quote) for job in jobs})
 
         # TODO: Shouldn't we return (CHF, USD) here, as above?
-        jobs = price.get_price_jobs_up_to_date(entries, datetime.date(2021, 1, 6),
-                                               False, None)
+        jobs = price.get_price_jobs_up_to_date(
+            entries, datetime.date(2021, 1, 6), False, None
+        )
         self.assertEqual(set(), {(job.base, job.quote) for job in jobs})
 
     @loader.load_doc()
@@ -582,19 +611,21 @@ class TestFilters(unittest.TestCase):
           Assets:US:Invest:Margin
         """
         jobs = price.get_price_jobs_up_to_date(entries, datetime.date(2021, 1, 8))
-        self.assertEqual({
-                ('QQQ', 'USD', datetime.date(2021, 1, 4)),
-                ('QQQ', 'USD', datetime.date(2021, 1, 5)),
-                ('QQQ', 'USD', datetime.date(2021, 1, 7)),
-                ('VEA', 'USD', datetime.date(2021, 1, 4)),
-                ('VEA', 'USD', datetime.date(2021, 1, 5)),
-                ('VEA', 'USD', datetime.date(2021, 1, 6)),
-                ('VEA', 'USD', datetime.date(2021, 1, 7)),
-            }, {(job.base, job.quote, job.date) for job in jobs})
+        self.assertEqual(
+            {
+                ("QQQ", "USD", datetime.date(2021, 1, 4)),
+                ("QQQ", "USD", datetime.date(2021, 1, 5)),
+                ("QQQ", "USD", datetime.date(2021, 1, 7)),
+                ("VEA", "USD", datetime.date(2021, 1, 4)),
+                ("VEA", "USD", datetime.date(2021, 1, 5)),
+                ("VEA", "USD", datetime.date(2021, 1, 6)),
+                ("VEA", "USD", datetime.date(2021, 1, 7)),
+            },
+            {(job.base, job.quote, job.date) for job in jobs},
+        )
 
 
 class TestFromFile(unittest.TestCase):
-
     @loader.load_doc()
     def setUp(self, entries, _, __):
         """
@@ -624,8 +655,8 @@ class TestFromFile(unittest.TestCase):
     def test_find_currencies_declared(self):
         currencies = price.find_currencies_declared(self.entries, None)
         currencies2 = [(base, quote) for base, quote, _ in currencies]
-        self.assertEqual([('QQQ', 'USD')], currencies2)
+        self.assertEqual([("QQQ", "USD")], currencies2)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
