@@ -18,33 +18,30 @@ def response(contents, status_code=requests.codes.ok):
     response.status_code = status_code
     response.text = ""
     response.json.return_value = contents
-    return mock.patch('requests.get', return_value=response)
+    return mock.patch("requests.get", return_value=response)
 
 
 class AlphavantagePriceFetcher(unittest.TestCase):
-
     def setUp(self):
-        environ['ALPHAVANTAGE_API_KEY'] = 'foo'
+        environ["ALPHAVANTAGE_API_KEY"] = "foo"
 
     def tearDown(self):
-        del environ['ALPHAVANTAGE_API_KEY']
+        del environ["ALPHAVANTAGE_API_KEY"]
 
     def test_error_invalid_ticker(self):
         with self.assertRaises(ValueError):
-            alphavantage.Source().get_latest_price('INVALID')
+            alphavantage.Source().get_latest_price("INVALID")
 
     def test_error_network(self):
-        with response('Foobar', 404):
+        with response("Foobar", 404):
             with self.assertRaises(alphavantage.AlphavantageApiError):
-                alphavantage.Source().get_latest_price('price:IBM:USD')
+                alphavantage.Source().get_latest_price("price:IBM:USD")
 
     def test_error_response(self):
-        contents = {
-            "Error Message": "Something wrong"
-        }
+        contents = {"Error Message": "Something wrong"}
         with response(contents):
             with self.assertRaises(alphavantage.AlphavantageApiError):
-                alphavantage.Source().get_latest_price('price:IBM:USD')
+                alphavantage.Source().get_latest_price("price:IBM:USD")
 
     def test_valid_response_price(self):
         contents = {
@@ -54,12 +51,13 @@ class AlphavantagePriceFetcher(unittest.TestCase):
             }
         }
         with response(contents):
-            srcprice = alphavantage.Source().get_latest_price('price:FOO:USD')
+            srcprice = alphavantage.Source().get_latest_price("price:FOO:USD")
             self.assertIsInstance(srcprice, source.SourcePrice)
-            self.assertEqual(Decimal('144.7400'), srcprice.price)
-            self.assertEqual('USD', srcprice.quote_currency)
-            self.assertEqual(datetime.datetime(2021, 1, 21, 0, 0, 0, tzinfo=tz.tzutc()),
-                             srcprice.time)
+            self.assertEqual(Decimal("144.7400"), srcprice.price)
+            self.assertEqual("USD", srcprice.quote_currency)
+            self.assertEqual(
+                datetime.datetime(2021, 1, 21, 0, 0, 0, tzinfo=tz.tzutc()), srcprice.time
+            )
 
     def test_valid_response_fx(self):
         contents = {
@@ -70,13 +68,14 @@ class AlphavantagePriceFetcher(unittest.TestCase):
             }
         }
         with response(contents):
-            srcprice = alphavantage.Source().get_latest_price('fx:USD:CHF')
+            srcprice = alphavantage.Source().get_latest_price("fx:USD:CHF")
             self.assertIsInstance(srcprice, source.SourcePrice)
-            self.assertEqual(Decimal('108.94000000'), srcprice.price)
-            self.assertEqual('CHF', srcprice.quote_currency)
-            self.assertEqual(datetime.datetime(2021, 2, 21, 20, 32, 25, tzinfo=tz.tzutc()),
-                             srcprice.time)
+            self.assertEqual(Decimal("108.94000000"), srcprice.price)
+            self.assertEqual("CHF", srcprice.quote_currency)
+            self.assertEqual(
+                datetime.datetime(2021, 2, 21, 20, 32, 25, tzinfo=tz.tzutc()), srcprice.time
+            )
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
