@@ -16,25 +16,24 @@ def response(contents, status_code=requests.codes.ok):
     response.status_code = status_code
     response.text = ""
     response.json.return_value = contents
-    return mock.patch('requests.get', return_value=response)
+    return mock.patch("requests.get", return_value=response)
 
 
 class CoinmarketcapPriceFetcher(unittest.TestCase):
-
     def setUp(self):
-        environ['COINMARKETCAP_API_KEY'] = 'foo'
+        environ["COINMARKETCAP_API_KEY"] = "foo"
 
     def tearDown(self):
-        del environ['COINMARKETCAP_API_KEY']
+        del environ["COINMARKETCAP_API_KEY"]
 
     def test_error_invalid_ticker(self):
         with self.assertRaises(ValueError):
-            coinmarketcap.Source().get_latest_price('INVALID')
+            coinmarketcap.Source().get_latest_price("INVALID")
 
     def test_error_network(self):
-        with response('Foobar', 404):
+        with response("Foobar", 404):
             with self.assertRaises(ValueError):
-                coinmarketcap.Source().get_latest_price('BTC-CHF')
+                coinmarketcap.Source().get_latest_price("BTC-CHF")
 
     def test_error_request(self):
         contents = {
@@ -45,7 +44,7 @@ class CoinmarketcapPriceFetcher(unittest.TestCase):
         }
         with response(contents):
             with self.assertRaises(ValueError):
-                coinmarketcap.Source().get_latest_price('BTC-CHF')
+                coinmarketcap.Source().get_latest_price("BTC-CHF")
 
     def test_valid_response(self):
         contents = {
@@ -54,7 +53,7 @@ class CoinmarketcapPriceFetcher(unittest.TestCase):
                     "quote": {
                         "CHF": {
                             "price": 1234.56,
-                            "last_updated": "2018-08-09T21:56:28.000Z"
+                            "last_updated": "2018-08-09T21:56:28.000Z",
                         }
                     }
                 }
@@ -62,14 +61,14 @@ class CoinmarketcapPriceFetcher(unittest.TestCase):
             "status": {
                 "error_code": 0,
                 "error_message": "",
-            }
+            },
         }
         with response(contents):
-            srcprice = coinmarketcap.Source().get_latest_price('BTC-CHF')
+            srcprice = coinmarketcap.Source().get_latest_price("BTC-CHF")
             self.assertIsInstance(srcprice, source.SourcePrice)
-            self.assertEqual(Decimal('1234.56'), srcprice.price)
-            self.assertEqual('CHF', srcprice.quote_currency)
+            self.assertEqual(Decimal("1234.56"), srcprice.price)
+            self.assertEqual("CHF", srcprice.quote_currency)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
