@@ -5,6 +5,7 @@ This is a really fantastic exchange API with a lot of relevant information.
 Timezone information: There is currency no support for historical prices. The
 output datetime is provided as a UNIX timestamp.
 """
+
 __copyright__ = "Copyright (C) 2018-2020  Martin Blais"
 __license__ = "GNU GPLv2"
 
@@ -27,24 +28,24 @@ def fetch_quote(ticker):
     url = "https://api.iextrading.com/1.0/tops/last?symbols={}".format(ticker.upper())
     response = requests.get(url)
     if response.status_code != requests.codes.ok:
-        raise IEXError("Invalid response ({}): {}".format(
-            response.status_code, response.text))
+        raise IEXError(
+            "Invalid response ({}): {}".format(response.status_code, response.text)
+        )
 
     results = response.json()
     if len(results) != 1:
-        raise IEXError("Invalid number of responses from IEX: {}".format(
-            response.text))
+        raise IEXError("Invalid number of responses from IEX: {}".format(response.text))
     result = results[0]
 
-    price = Decimal(result['price']).quantize(Decimal('0.01'))
+    price = Decimal(result["price"]).quantize(Decimal("0.01"))
 
     # IEX is American markets.
     us_timezone = tz.gettz("America/New_York")
-    time = datetime.datetime.fromtimestamp(result['time'] / 1000)
+    time = datetime.datetime.fromtimestamp(result["time"] / 1000)
     time = time.astimezone(us_timezone)
 
     # As far as can tell, all the instruments on IEX are priced in USD.
-    return source.SourcePrice(price, time, 'USD')
+    return source.SourcePrice(price, time, "USD")
 
 
 class Source(source.Source):
@@ -58,4 +59,5 @@ class Source(source.Source):
         """See contract in beanprice.source.Source."""
         raise NotImplementedError(
             "This is now implemented at https://iextrading.com/developers/docs/#hist and "
-            "needs to be added here.")
+            "needs to be added here."
+        )
