@@ -1,4 +1,4 @@
-""" Fetch prices from US Government Thrift Savings Plan
+"""Fetch prices from US Government Thrift Savings Plan
 
 As of 7 July 2020, the Thrift Savings Plan (TSP) rolled out a new
 web site that has an API (instead of scraping a CSV). Unable to
@@ -7,6 +7,7 @@ is available at:
 
 https://secure.tsp.gov/components/CORS/
 """
+
 __copyright__ = "Copyright (C) 2020 Martin Blais"
 __license__ = "GNU GPLv2"
 
@@ -20,39 +21,44 @@ import requests
 from beanprice import source
 
 # All of the TSP funds are in USD.
-CURRENCY = 'USD'
+CURRENCY = "USD"
 
-TIMEZONE = datetime.timezone(datetime.timedelta(hours=-4), 'America/New_York')
+TIMEZONE = datetime.timezone(datetime.timedelta(hours=-4), "America/New_York")
 
 TSP_FUND_NAMES = [
-    "LInco", #0
-    "L2025", #1
-    "L2030", #2
-    "L2035", #3
-    "L2040", #4
-    "L2045", #5
-    "L2050", #6
-    "L2055", #7
-    "L2060", #8
-    "L2065", #9
-    "GFund", #10
-    "FFund", #11
-    "CFund", #12
-    "SFund", #13
-    "IFund", #14
+    "LInco",  # 0
+    "L2025",  # 1
+    "L2030",  # 2
+    "L2035",  # 3
+    "L2040",  # 4
+    "L2045",  # 5
+    "L2050",  # 6
+    "L2055",  # 7
+    "L2060",  # 8
+    "L2065",  # 9
+    "GFund",  # 10
+    "FFund",  # 11
+    "CFund",  # 12
+    "SFund",  # 13
+    "IFund",  # 14
 ]
 
-csv.register_dialect('tsp',
-                     delimiter=',',
-                     quoting=csv.QUOTE_NONE,
-                     quotechar='',
-                     lineterminator='\n')
+csv.register_dialect(
+    "tsp",
+    delimiter=",",
+    quoting=csv.QUOTE_NONE,
+    # NOTE(blais): This fails to import in 3.12 (and perhaps before).
+    # quotechar='',
+    lineterminator="\n",
+)
+
 
 class TSPError(ValueError):
     "An error from the Thrift Savings Plan (TSP) API."
 
+
 def parse_tsp_csv(response: requests.models.Response) -> OrderedDict:
-    """ Parses a Thrift Savings Plan output CSV file.
+    """Parses a Thrift Savings Plan output CSV file.
 
     Function takes in a requests response and returns an
     OrderedDict with newest closing cost at front of OrderedDict.
@@ -62,32 +68,36 @@ def parse_tsp_csv(response: requests.models.Response) -> OrderedDict:
 
     text = response.iter_lines(decode_unicode=True)
 
-    reader = csv.DictReader(text, dialect='tsp')
+    reader = csv.DictReader(text, dialect="tsp")
 
     for row in reader:
         # Date from TSP looks like "July 30. 2020"
         # There is indeed a period after the day of month.
-        date = datetime.datetime.strptime(row['Date'], "%b %d. %Y")
+        date = datetime.datetime.strptime(row["Date"], "%b %d. %Y")
         date = date.replace(hour=16, tzinfo=TIMEZONE)
-        names = ['L Income',
-                 'L 2025',
-                 'L 2030',
-                 'L 2035',
-                 'L 2040',
-                 'L 2045',
-                 'L 2050',
-                 'L 2055',
-                 'L 2060',
-                 'L 2065',
-                 'G Fund',
-                 'F Fund',
-                 'C Fund',
-                 'S Fund',
-                 'I Fund']
-        data[date] = [Decimal(row[name]) if row[name] else Decimal()
-                      for name in map(str.strip, names)]
+        names = [
+            "L Income",
+            "L 2025",
+            "L 2030",
+            "L 2035",
+            "L 2040",
+            "L 2045",
+            "L 2050",
+            "L 2055",
+            "L 2060",
+            "L 2065",
+            "G Fund",
+            "F Fund",
+            "C Fund",
+            "S Fund",
+            "I Fund",
+        ]
+        data[date] = [
+            Decimal(row[name]) if row[name] else Decimal() for name in map(str.strip, names)
+        ]
 
     return OrderedDict(sorted(data.items(), key=lambda t: t[0], reverse=True))
+
 
 def parse_response(response: requests.models.Response) -> OrderedDict:
     """Process as response from TSP.
@@ -116,23 +126,23 @@ class Source(source.Source):
         if fund not in TSP_FUND_NAMES:
             raise TSPError(
                 "Invalid TSP Fund Name '{}'. Valid Funds are:\n\t{}".format(
-                    fund,
-                    "\n\t".join(TSP_FUND_NAMES)))
+                    fund, "\n\t".join(TSP_FUND_NAMES)
+                )
+            )
 
         url = "https://secure.tsp.gov/components/CORS/getSharePricesRaw.html"
-        fields = ['startdate', 'enddate', 'download', 'Lfunds', 'InvFunds']
         payload = {
             # Grabbing the last fourteen days of data in event the markets were closed.
-            'startdate' : (time - datetime.timedelta(days=14)).strftime("%Y%m%d"),
-            'enddate':  time.strftime("%Y%m%d"),
-            'download': '0',
-            'Lfunds': '1',
-            'InvFunds' : '1'
+            "startdate": (time - datetime.timedelta(days=14)).strftime("%Y%m%d"),
+            "enddate": time.strftime("%Y%m%d"),
+            "download": "0",
+            "Lfunds": "1",
+            "InvFunds": "1",
         }
 
         response = requests.get(url, params=payload)
         result = parse_response(response)
-        trade_day = list(result.items())[0]
+        trade_day = next(iter(result.items()))
         prices = trade_day[1]
 
         try:
