@@ -27,9 +27,9 @@ class MockResponse:
 
 
 class YahooFinancePriceFetcher(unittest.TestCase):
-
     def _test_get_latest_price(self):
-        response = MockResponse(textwrap.dedent("""
+        response = MockResponse(
+            textwrap.dedent("""
             {"quoteResponse":
              {"error": null,
               "result": [{"esgPopulated": false,
@@ -48,15 +48,18 @@ class YahooFinancePriceFetcher(unittest.TestCase):
                           "sourceInterval": 15,
                           "symbol": "XSP.TO",
                           "tradeable": false}]}}
-            """))
-        with mock.patch('requests.get', return_value=response):
-            srcprice = yahoo.Source().get_latest_price('XSP.TO')
+            """)
+        )
+        yahoo_source = yahoo.Source()
+        with mock.patch.object(yahoo_source.session, "get", return_value=response):
+            srcprice = yahoo_source.get_latest_price("XSP.TO")
         self.assertTrue(isinstance(srcprice.price, Decimal))
-        self.assertEqual(Decimal('29.99'), srcprice.price)
-        timezone = datetime.timezone(datetime.timedelta(hours=-4), 'America/Toronto')
-        self.assertEqual(datetime.datetime(2018, 3, 29, 15, 59, 49, tzinfo=timezone),
-                         srcprice.time)
-        self.assertEqual('CAD', srcprice.quote_currency)
+        self.assertEqual(Decimal("29.99"), srcprice.price)
+        timezone = datetime.timezone(datetime.timedelta(hours=-4), "America/Toronto")
+        self.assertEqual(
+            datetime.datetime(2018, 3, 29, 15, 59, 49, tzinfo=timezone), srcprice.time
+        )
+        self.assertEqual("CAD", srcprice.quote_currency)
 
     def test_get_latest_price(self):
         for tzname in "America/New_York", "Europe/Berlin", "Asia/Tokyo":
@@ -64,7 +67,8 @@ class YahooFinancePriceFetcher(unittest.TestCase):
                 self._test_get_latest_price()
 
     def _test_get_historical_price(self):
-        response = MockResponse(textwrap.dedent("""
+        response = MockResponse(
+            textwrap.dedent("""
             {"chart":
              {"error": null,
               "result": [{"indicators": {"adjclose": [{"adjclose": [29.236251831054688,
@@ -118,16 +122,20 @@ class YahooFinancePriceFetcher(unittest.TestCase):
                           "timestamp": [1509111000,
                                         1509370200,
                                         1509456600,
-                                        1509543000]}]}}"""))
-        with mock.patch('requests.get', return_value=response):
-            srcprice = yahoo.Source().get_historical_price(
-                'XSP.TO', datetime.datetime(2017, 11, 1, 16, 0, 0, tzinfo=tz.tzutc()))
+                                        1509543000]}]}}""")
+        )
+        yahoo_source = yahoo.Source()
+        with mock.patch.object(yahoo_source.session, "get", return_value=response):
+            srcprice = yahoo_source.get_historical_price(
+                "XSP.TO", datetime.datetime(2017, 11, 1, 16, 0, 0, tzinfo=tz.tzutc())
+            )
         self.assertTrue(isinstance(srcprice.price, Decimal))
-        self.assertEqual(Decimal('29.469999313354492'), srcprice.price)
-        timezone = datetime.timezone(datetime.timedelta(hours=-4), 'America/Toronto')
-        self.assertEqual(datetime.datetime(2017, 11, 1, 9, 30, tzinfo=timezone),
-                         srcprice.time)
-        self.assertEqual('CAD', srcprice.quote_currency)
+        self.assertEqual(Decimal("29.469999313354492"), srcprice.price)
+        timezone = datetime.timezone(datetime.timedelta(hours=-4), "America/Toronto")
+        self.assertEqual(
+            datetime.datetime(2017, 11, 1, 9, 30, tzinfo=timezone), srcprice.time
+        )
+        self.assertEqual("CAD", srcprice.quote_currency)
 
     def test_get_historical_price(self):
         for tzname in "America/New_York", "Europe/Berlin", "Asia/Tokyo":
@@ -136,32 +144,34 @@ class YahooFinancePriceFetcher(unittest.TestCase):
 
     def test_parse_response_error_status_code(self):
         response = MockResponse(
-            '{"quoteResponse": {"error": "Not supported", "result": [{}]}}',
-            status_code=400)
+            '{"quoteResponse": {"error": "Not supported", "result": [{}]}}', status_code=400
+        )
         with self.assertRaises(yahoo.YahooError):
             yahoo.parse_response(response)
 
     def test_parse_response_error_invalid_format(self):
         response = MockResponse(
             """{"quoteResponse": {"error": null, "result": [{}]},
-             "chart": {"error": null, "result": [{}]}}""")
+             "chart": {"error": null, "result": [{}]}}"""
+        )
         with self.assertRaises(yahoo.YahooError):
             yahoo.parse_response(response)
 
     def test_parse_response_error_not_none(self):
         response = MockResponse(
-            '{"quoteResponse": {"error": "Non-zero error", "result": [{}]}}')
+            '{"quoteResponse": {"error": "Non-zero error", "result": [{}]}}'
+        )
         with self.assertRaises(yahoo.YahooError):
             yahoo.parse_response(response)
 
     def test_parse_response_empty_result(self):
-        response = MockResponse(
-            '{"quoteResponse": {"error": null, "result": []}}')
+        response = MockResponse('{"quoteResponse": {"error": null, "result": []}}')
         with self.assertRaises(yahoo.YahooError):
             yahoo.parse_response(response)
 
     def test_parse_response_no_timestamp(self):
-        response = MockResponse(textwrap.dedent("""
+        response = MockResponse(
+            textwrap.dedent("""
             {"chart":
              {"error": null,
               "result": [{"indicators": {"adjclose": [{}],
@@ -190,15 +200,18 @@ class YahooFinancePriceFetcher(unittest.TestCase):
                                    "timezone": "EDT",
                                    "validRanges": ["1d", "5d", "1mo", "3mo", "6mo", "1y",
                                                    "2y", "5y", "10y", "ytd", "max"]}}]}}
-        """))
+        """)
+        )
         with self.assertRaises(yahoo.YahooError):
-            with mock.patch('requests.get', return_value=response):
-                srcprice = yahoo.Source().get_historical_price(
-                    'XSP.TO', datetime.datetime(2017, 11, 1, 16, 0, 0, tzinfo=tz.tzutc()))
-
+            yahoo_source = yahoo.Source()
+            with mock.patch.object(yahoo_source.session, "get", return_value=response):
+                _ = yahoo_source.get_historical_price(
+                    "XSP.TO", datetime.datetime(2017, 11, 1, 16, 0, 0, tzinfo=tz.tzutc())
+                )
 
     def test_parse_null_prices_in_series(self):
-        response = MockResponse(textwrap.dedent("""
+        response = MockResponse(
+            textwrap.dedent("""
             {"chart": {"result":[ {"meta":{
                 "currency":"USD","symbol":"FBIIX",
                 "exchangeName":"NAS","instrumentType":"MUTUALFUND",
@@ -226,16 +239,22 @@ class YahooFinancePriceFetcher(unittest.TestCase):
                         {"adjclose":[9.6899995803833,9.710000038146973,9.6899995803833,null]}
                     ]
                 }}],"error":null}}
-        """))
+        """)
+        )
 
-        with mock.patch('requests.get', return_value=response):
-            srcprice = yahoo.Source().get_historical_price(
-                'XSP.TO', datetime.datetime(2022, 2, 28, 16, 0, 0, tzinfo=tz.tzutc()))              
+        yahoo_source = yahoo.Source()
+        with mock.patch.object(yahoo_source.session, "get", return_value=response):
+            srcprice = yahoo_source.get_historical_price(
+                "XSP.TO", datetime.datetime(2022, 2, 28, 16, 0, 0, tzinfo=tz.tzutc())
+            )
             self.assertTrue(isinstance(srcprice.price, Decimal))
-            self.assertEqual(Decimal('9.6899995803833'), srcprice.price)
-            timezone = datetime.timezone(datetime.timedelta(hours=-5), 'America/New_York')
-            self.assertEqual(datetime.datetime(2022, 2, 25, 9, 30, tzinfo=timezone),
-                            srcprice.time)
-            self.assertEqual('USD', srcprice.quote_currency)
-if __name__ == '__main__':
+            self.assertEqual(Decimal("9.6899995803833"), srcprice.price)
+            timezone = datetime.timezone(datetime.timedelta(hours=-5), "America/New_York")
+            self.assertEqual(
+                datetime.datetime(2022, 2, 25, 9, 30, tzinfo=timezone), srcprice.time
+            )
+            self.assertEqual("USD", srcprice.quote_currency)
+
+
+if __name__ == "__main__":
     unittest.main()
